fix: correct strncmp_k integer arithmetic and clarify stub function comments
Agent-Logs-Url: https://github.com/Cbetts1/NEW-ATTEMPT/sessions/e04e9e19-d795-4676-b87e-9f148dfe7db6

+++ kernel/plugin.c
@@ -14,7 +14,7 @@ static int           plugin_count_val = 0;
 
 static int strncmp_k(const char *a, const char *b, int n) {
     while (n-- > 0) {
-        if (*a != *b) return (unsigned char)*a - (unsigned char)*b;
+        if (*a != *b) return (int)(unsigned char)*a - (int)(unsigned char)*b;
         if (*a == '\0') return 0;
         a++; b++;
     }
